Revert "Move operator lifecycle under internal (#1691)"
This reverts commit 4e5fa7cab81b3e3cae2b700bd272672f6bdb9db2.
