+++ gestaltd/internal/daemon/factories.go
@@ -13,11 +13,11 @@ import (
 	coreworkflow "github.com/valon-technologies/gestalt/server/core/workflow"
 	"github.com/valon-technologies/gestalt/server/internal/bootstrap"
 	"github.com/valon-technologies/gestalt/server/internal/config"
-	"github.com/valon-technologies/gestalt/server/internal/operator"
 	"github.com/valon-technologies/gestalt/server/services/invocation"
 	telemetrynoop "github.com/valon-technologies/gestalt/server/services/observability/drivers/noop"
 	telemetryotlp "github.com/valon-technologies/gestalt/server/services/observability/drivers/otlp"
 	telemetrystdout "github.com/valon-technologies/gestalt/server/services/observability/drivers/stdout"
+	"github.com/valon-technologies/gestalt/server/services/operator"
 	"github.com/valon-technologies/gestalt/server/services/providerdrivers"
 	"github.com/valon-technologies/gestalt/server/services/runtimehost"
 	secretsenv "github.com/valon-technologies/gestalt/server/services/secrets/drivers/env"
