+++ gestaltd/services/operator/lifecycle_test.go
