+++ gestaltd/services/operator/localconfig_test.go
