+++ gestaltd/services/operator/testmain_test.go
